Docker image to host cSpot
Created a Dockerfile for building a Docker image which can host cSpot.
It includes the dependencies required by the cSpot project, not all of
these were apparent in the project documentation - particularly a (seeming) error in the naming of a
app directory, app/, which was referenced in the code as App/ instead.
Since it took me some time to work out how to install the project
correctly, I thought it would be helpful to write this Dockerfile to
simplify the process for other users/developers. Plus, Docker is awesome :)

I've included some instructions in the readme, to get started with
running the docker container.